% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glinvci.R
\docType{package}
\name{glinvci}
\alias{glinvci}
\alias{glinvci-package}
\title{glinvci: Confidence intervals and hypothesis testing for GLInv model}
\description{
The glinvci package provides a framework for computing the maximum-likelihood estimates
and asymptotic confidence intervals of a class of continuous-time Gaussian branching processes,
including the Ornstein-Uhlenbeck branching process, which is commonly used in phylogenetic
comparative methods. The framework is designed to be flexible enough that the user can
easily specify their own parameterisation and obtain the maximum-likelihood estimates and
confidence intervals of their own parameters.
}
\author{
Hao Chi Kiang, \email{hello@hckiang.com}
}
