% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glinvci.R
\name{grad}
\alias{grad}
\title{Compute the log-likelihood gradients of GLInv models}
\usage{
grad(mod, ...)
}
\arguments{
\item{mod}{An object of either \code{\link{glinv}} or \code{\link{glinv_gauss}} class.}

\item{...}{Further arguments to be passed to the S3 methods.}
}
\value{
A numerical vector containing the gradient of \code{mod}.
}
\description{
For the \code{\link{glinv}} class, which is a high-level user interface, please see \code{\link{grad.glinv}}; and
for \code{\link{glinv_gauss}}, which is a lower-level facility, please see \code{\link{grad.glinv_gauss}}.
}
