% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glinvci.R
\name{has_tipvals}
\alias{has_tipvals}
\alias{has_tipvals.glinv_gauss}
\title{Check if a \code{glinv_gauss} model contains trait values at their tips.}
\usage{
has_tipvals(mod)

\method{has_tipvals}{glinv_gauss}(mod)
}
\arguments{
\item{mod}{A \code{glinv_gauss} object.}
}
\description{
Returns true if and only if the \code{glinv_gauss} model were initialised with \code{X=NULL}
and the user had never called \code{set_tips} on it.
}
