% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glinvci.R
\name{hess}
\alias{hess}
\title{Compute the log-likelihood Hessian of GLInv models}
\usage{
hess(mod, ...)
}
\arguments{
\item{mod}{An object of either \code{\link{glinv}} or \code{\link{glinv_gauss}} class.}

\item{...}{Further arguments to be passed to the S3 methods.}
}
\value{
A numerical square matrix containing the Hessian of \code{mod}.
}
\description{
For the \code{\link{glinv}} class, which is a high-level user interface, please see \code{\link{hess.glinv}}; and
for \code{\link{glinv_gauss}}, which is a lower-level facility, please see \code{\link{hess.glinv_gauss}}.
}
