% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glinvci.R
\name{lik}
\alias{lik}
\title{Compute the likelihood of a GLInv model}
\usage{
lik(mod, ...)
}
\arguments{
\item{mod}{An object of either \code{\link{glinv}} or \code{\link{glinv_gauss}} class.}

\item{...}{Further arguments to be passed to the S3 methods.}
}
\description{
This is a S3 generic method. For the \code{\link{glinv}} class, which is a high-level user interface, please 
see \code{\link{lik.glinv}}; and for \code{\link{glinv_gauss}}, which is a lower-level facility, please see
\code{\link{lik.glinv_gauss}}.
}
