% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glinvci.R
\name{marginal_ci}
\alias{marginal_ci}
\title{Getting marginal confidence interval for GLInv model}
\usage{
marginal_ci(varest_result, lvl = 0.95)
}
\arguments{
\item{varest_result}{The output from `varest.glinv`.}

\item{lvl}{Confidence level. Default to 95 percent.}
}
\value{
A matrix \eqn{p}-by-2 matrix where \eqn{p} is the number of parameters.
                        The first column is the lower limits and second column is the upper limits.
}
\description{
\code{marginal_ci} computes the marginal confidence interval for each parameters
using the variance-covariance matrix output by `varest.glinv`.
}
