% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ou.R
\name{nparams_ou}
\alias{nparams_ou}
\title{Get the number of parameters of some pre-defined models}
\usage{
nparams_ou(k)
}
\arguments{
\item{k}{An Integer. The total number of dimensions of the multivariate traits.}
}
\description{
\code{nparams_ou} returns the number of parameters of their respective model.
}
