% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glinvci.R
\name{rglinv}
\alias{rglinv}
\alias{rglinv.glinv}
\alias{rglinv.glinv_gauss}
\title{Simulate random trait values from models.}
\usage{
rglinv(mod, par, Nsamp)

\method{rglinv}{glinv}(mod, par, Nsamp)

\method{rglinv}{glinv_gauss}(mod, par, Nsamp)
}
\arguments{
\item{mod}{Either a \code{glinv_gauss} or \code{glinv} object.}

\item{par}{Parameters underlying the simulation, in the same format as \code{lik.glinv_gauss} or \code{lik.glinv}.}

\item{Nsamp}{Number of sample point to simulate.}
}
\value{
A list containing Nsamp elements, each of which is a length-n list, where \eqn{n} is the number of tips
              containing the simulated trait values.
}
\description{
Simulate random trait values from the Gaussian branching process specified by \code{mod}.
}
